Use arguments to pass Ollama API endpoint and model. Add support for JPEG formatted images.
This commit modifies the main.ts file to use arguments to pass the Ollama API endpoint and model as command line arguments. Previously, these were hardcoded into the file. This change allows for more flexibility when running the script, as the API endpoint and model can now be specified each time the script is run. Additionally, a new directory called ai-renamed is created if it doesn't exist already and renamed images are created there. Also support for JPEG images is added.