Replace node fs with deno standard library fs for better interoperability and to avoid external dependency on node fs module
This commit replaces the existing import of the fs module from node with the equivalent module from the Deno standard library. The reasoning behind this change is to make the code work in a Deno environment, since the fs module from node is not available there. This will also ensure better interoperability between Deno projects as it removes an external dependency on the node fs module